requirements.txt
transitions
pywin32
pywin32-ctypes
pyautogui
opencv-python
pydirectinput
matplotlib
keyboard
hidapi